Reworked MultiplayerServer.js to include app.js script stuff
